requirements.txt
django-payments>=0.12.0
